Day 24 - Never Tell Me The Odds - part 2 - more generic equation used to cover vertical lines

+++ src/main/java/advent/day24/Intersection.java
@@ -3,9 +3,9 @@ package advent.day24;
 import java.math.BigDecimal;
 
 public class Intersection {
-    double x, y;
+    BigDecimal x, y;
 
-    public Intersection(double x, double y) {
+    public Intersection(BigDecimal x, BigDecimal y) {
         this.x = x;
         this.y = y;
     }
